feat(products by category): implement view product by category  - fix backlog order bug  [Finishes #153209570]
