+++ imports/plugins/included/product-variant/containers/gridItemNoticeContainer.js
@@ -34,10 +34,8 @@ const wrapComponent = (Comp) => (
 
     isSoldOut = () => {
       const topVariants = ReactionProduct.getTopVariants(this.props.product._id);
-
       for (const topVariant of topVariants) {
         const inventoryQuantity = ReactionProduct.getVariantQuantity(topVariant);
-
         if (inventoryQuantity > 0) {
           return false;
         }
